moved HookException and UnsupportedVersionException into compatibility package
