+++ src/main/java/org/betonquest/betonquest/compatibility/holograms/HologramIntegrator.java
@@ -2,9 +2,9 @@ package org.betonquest.betonquest.compatibility.holograms;
 
 import org.betonquest.betonquest.BetonQuest;
 import org.betonquest.betonquest.api.config.quest.QuestPackage;
+import org.betonquest.betonquest.compatibility.HookException;
 import org.betonquest.betonquest.compatibility.Integrator;
-import org.betonquest.betonquest.exception.HookException;
-import org.betonquest.betonquest.exception.UnsupportedVersionException;
+import org.betonquest.betonquest.compatibility.UnsupportedVersionException;
 import org.betonquest.betonquest.versioning.UpdateStrategy;
 import org.betonquest.betonquest.versioning.Version;
 import org.betonquest.betonquest.versioning.VersionComparator;
