+++ src/main/java/org/betonquest/betonquest/compatibility/holograms/decentholograms/DecentHologramsIntegrator.java
@@ -7,10 +7,10 @@ import org.betonquest.betonquest.api.Variable;
 import org.betonquest.betonquest.api.config.quest.QuestPackage;
 import org.betonquest.betonquest.api.logger.BetonQuestLogger;
 import org.betonquest.betonquest.api.quest.QuestException;
+import org.betonquest.betonquest.compatibility.HookException;
 import org.betonquest.betonquest.compatibility.holograms.BetonHologram;
 import org.betonquest.betonquest.compatibility.holograms.HologramIntegrator;
 import org.betonquest.betonquest.compatibility.holograms.HologramProvider;
-import org.betonquest.betonquest.exception.HookException;
 import org.betonquest.betonquest.instruction.Instruction;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
