+++ src/main/java/org/betonquest/betonquest/compatibility/luckperms/LuckPermsIntegrator.java
@@ -3,9 +3,9 @@ package org.betonquest.betonquest.compatibility.luckperms;
 import net.luckperms.api.LuckPerms;
 import net.luckperms.api.context.ContextCalculator;
 import org.betonquest.betonquest.BetonQuest;
+import org.betonquest.betonquest.compatibility.HookException;
 import org.betonquest.betonquest.compatibility.Integrator;
 import org.betonquest.betonquest.compatibility.luckperms.permission.LuckPermsEventFactory;
-import org.betonquest.betonquest.exception.HookException;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.RegisteredServiceProvider;
