+++ src/main/java/org/betonquest/betonquest/compatibility/protocollib/ProtocolLibIntegrator.java
@@ -1,11 +1,11 @@
 package org.betonquest.betonquest.compatibility.protocollib;
 
 import org.betonquest.betonquest.BetonQuest;
+import org.betonquest.betonquest.compatibility.HookException;
 import org.betonquest.betonquest.compatibility.Integrator;
+import org.betonquest.betonquest.compatibility.UnsupportedVersionException;
 import org.betonquest.betonquest.compatibility.protocollib.conversation.MenuConvIO;
 import org.betonquest.betonquest.compatibility.protocollib.conversation.PacketInterceptor;
-import org.betonquest.betonquest.exception.HookException;
-import org.betonquest.betonquest.exception.UnsupportedVersionException;
 import org.betonquest.betonquest.versioning.UpdateStrategy;
 import org.betonquest.betonquest.versioning.Version;
 import org.betonquest.betonquest.versioning.VersionComparator;
