+++ src/main/java/org/betonquest/betonquest/compatibility/redischat/RedisChatIntegrator.java
@@ -1,8 +1,8 @@
 package org.betonquest.betonquest.compatibility.redischat;
 
 import org.betonquest.betonquest.BetonQuest;
+import org.betonquest.betonquest.compatibility.HookException;
 import org.betonquest.betonquest.compatibility.Integrator;
-import org.betonquest.betonquest.exception.HookException;
 import org.bukkit.event.Listener;
 
 /**
