+++ src/main/java/org/betonquest/betonquest/compatibility/traincarts/TrainCartsIntegrator.java
@@ -1,12 +1,12 @@
 package org.betonquest.betonquest.compatibility.traincarts;
 
 import org.betonquest.betonquest.BetonQuest;
+import org.betonquest.betonquest.compatibility.HookException;
 import org.betonquest.betonquest.compatibility.Integrator;
 import org.betonquest.betonquest.compatibility.traincarts.conditions.TrainCartsRideConditionFactory;
 import org.betonquest.betonquest.compatibility.traincarts.objectives.TrainCartsExitObjective;
 import org.betonquest.betonquest.compatibility.traincarts.objectives.TrainCartsLocationObjective;
 import org.betonquest.betonquest.compatibility.traincarts.objectives.TrainCartsRideObjective;
-import org.betonquest.betonquest.exception.HookException;
 import org.betonquest.betonquest.quest.PrimaryServerThreadData;
 import org.betonquest.betonquest.quest.registry.type.ObjectiveTypeRegistry;
